requirements/tests.txt
pytest==3.4.0
pytest-variables==1.7.1
pytest-selenium==1.11.4
pyotp==2.2.6
pytest-xdist==1.22.0